add connectivity service

+++ lib/model/getproduct_model/getproduct_model.dart
@@ -49,7 +49,7 @@ class Rating {
   Rating({this.rate,count});
 
   Rating.fromJson(Map<String, dynamic> json) {
-    rate = json['rate'];
+    rate = json['rate'].toDouble();
     count = json['count'];
   }
